restructure files into a package
