+++ gztprocessor/gazette_processors/mindep_gazette_processor.py
@@ -1,14 +1,12 @@
 from pathlib import Path
 import re
 from collections import defaultdict
-from db_connections.db_gov import get_connection
-from state_managers.mindep_state_manager import MindepStateManager
-import utils
+from gztprocessor.db_connections.db_gov import get_connection
+from gztprocessor.state_managers.mindep_state_manager import MindepStateManager
 
 mindep_state_manager = MindepStateManager()
 
-def extract_initial_gazette_data(gazette_number: str, date_str: str) -> dict:
-    data = utils.load_mindep_gazette_data_from_JSON(gazette_number, date_str)
+def extract_initial_gazette_data(gazette_number: str, date_str: str, data: dict) -> dict:
     ministries = data.get("ministers", [])
     if not ministries:
         raise ValueError(
@@ -18,8 +16,7 @@ def extract_initial_gazette_data(gazette_number: str, date_str: str) -> dict:
 
 # TODO: Resolve this issue for renames: https://github.com/zaeema-n/orgchart_nexoan/issues/11#issue-3238949430
 
-def extract_column_II_department_changes(gazette_number: str, date_str: str) -> tuple[list[dict], list[dict]]:
-    data = utils.load_mindep_gazette_data_from_JSON(gazette_number, date_str)
+def extract_column_II_department_changes(gazette_number: str, date_str: str, data: dict) -> tuple[list[dict], list[dict]]:
 
     adds = [e for e in data.get("ADD", []) if e.get("affected_column") == "II"]
     omits = [e for e in data.get("OMIT", []) if e.get("affected_column") == "II"]
