+++ gztprocessor/schemas/mindep_schema.sql
