+++ gztprocessor/state_managers/state_manager.py
