+++ routes/person_router.py
@@ -3,11 +3,12 @@ from fastapi import APIRouter, Body
 from fastapi.params import Body
 from typing import List
 
-from state_managers.person_state_manager import PersonStateManager
-import gazette_processors.person_gazette_processor as person_gazette_processor
-import database_handlers.person_database_handler as person_database
-import csv_writer
+from gztprocessor.state_managers.person_state_manager import PersonStateManager
+import gztprocessor.gazette_processors.person_gazette_processor as person_gazette_processor
+import gztprocessor.database_handlers.person_database_handler as person_database
+import gztprocessor.csv_writer as csv_writer
 from routes.state_router import create_state_routes
+import utils as utils
 
 person_router = APIRouter()
 person_state_manager = PersonStateManager()
@@ -20,7 +21,8 @@ def get_contents_of_person_gazette(gazette_number: str, date: str):
     Return the predicted transactions from a person gazette.
     """
     try:
-        transactions = person_gazette_processor.process_person_gazette(gazette_number, date)
+        data = utils.load_person_gazette_data_from_JSON(gazette_number, date)
+        transactions = person_gazette_processor.process_person_gazette(gazette_number, date, data)
         return transactions
     except FileNotFoundError:
         return {"error": f"Gazette file for {gazette_number}, {date} not found."}
